Begin to add types `quintptr`, `qintptr`, `QDate`, `QTime`, `QDateTime`, `QSize`, `QSizeF`, `QPoint`, `QPointF`, `QRect`, `QRectF` to the interface.
On the Qt-side, these types are represented by the qt-type with the same name. On the rust-side, the date/time-types are represented by the "naive" types of the "chrono"-crate, the geometric types are represented as simple tuples and the pointer-sized types are represented by libc's `uintptr_t` and `intptr_t`.

As most of these types are "complex" (similar to `QString` and `QByteArray`), they have to be converted by special functions on the ffi-boundary or be passed by special function signatures (like taking two integers `x` and `y` instead of one value for a `QPoint`).

This commit handles these types in the interface.rs file for simple properties. Itemproperties of models and the template for the implementation.rs file are not yet finished.